Moved file managers from DynamicItems to separate module.
