+++ FrontEnd/Modules/FileManager/Enums/FileManagerModes.cs
@@ -1,0 +1,8 @@
+﻿namespace FrontEnd.Modules.FileManager.Enums;
+
+public enum FileManagerModes
+{
+    Images,
+    Files,
+    Templates
+}
