+++ FrontEnd/Modules/FileManager/Models/FileManagerViewModel.cs
@@ -1,0 +1,9 @@
+﻿using FrontEnd.Modules.Base.Models;
+using FrontEnd.Modules.FileManager.Enums;
+
+namespace FrontEnd.Modules.FileManager.Models;
+
+public class FileManagerViewModel : BaseModuleViewModel
+{
+    public FileManagerModes? Mode { get; set; }
+}
