+++ FrontEnd/Modules/Templates/Css/Templates.css
@@ -551,63 +551,6 @@
     height: 700px
 }
 
-/** ICONS **/
-.k-sprite {
-    color: var(--dark-txt);
-    font-family: 'icomoon' !important;
-    font-size: 16px;
-    margin-right: 5px;
-    padding-top: 1px;
-    speak: none;
-    font-style: normal;
-    font-weight: normal;
-    font-variant: normal;
-    text-transform: none;
-    line-height: 1;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-}
-
-.k-sprite.rootfolder:before {
-    content: "\e972";
-}
-
-.k-sprite.folder:before {
-    content: "\e968";
-}
-
-.k-sprite.wiserfolderclosed:before {
-    content: "\e96d";
-}
-
-.k-sprite.wiserfolderopened:before {
-    content: "\e965";
-}
-
-.k-sprite.pdf:before {
-    content: "\e956";
-}
-
-.k-sprite.html:before {
-    content: "\e957";
-}
-
-.k-sprite.image:before {
-    content: "\e9eb";
-}
-
-.k-sprite.file:before {
-    content: "\e94b";
-}
-
-.k-sprite.edit:before {
-    content: "\e9e8";
-}
-
-.k-sprite.delete:before {
-    content: "\e93f";
-}
-
 .k-state-default > .k-clear-value {
     position: absolute;
     top: 0;
